Refactoring server
